feat: reverted standalone change made and moved the helpers to list secrets
